Use cloud name from configuration
If we specify multiple clouds for example into clouds.yaml file, we need
to have a way how to pick which cloud to use for specific backup. For
this we can set cloud name into BackupStorageLocation or
VolumeSnapshotLocation and plugin will pick which cloud to use for
authentication and further operations.

Signed-off-by: Ondrej Vasko <ondrej.vaskoo@gmail.com>