Revert: set default region to empty string ""
For non-documented reason we switched from having default value for
region empty string to value RegionOne (in tag v0.4.0). RegionOne is default region name
in Openstack, but this doesn't justify leaving region to default to this
value.

Having region as empty string makes the gophercloud library to pick
first found region in the catalog. This can cause potential backwards
compatibility issues, which are however very unlikely.

Negative impact of having region default to RegionOne is that it makes
the plugin to not work in case user relies on region specified in
clouds.yaml. This also looks slightly tricky as region in clouds.yaml
doesn't neccesarilly mean that that region should be configured in
Cinder.

This change and additional test displays the behavior.

Signed-off-by: Ondrej Vasko <ondrej.vaskoo@gmail.com>