Set default region back to empty string (#110)
For all plugins - block, file share and object storage the default region name 
will be `""` and plugin will respect setting of `region_name` configured in clouds.yaml.

For non-documented reason we switched from having default value for
region to be empty string to be value `RegionOne` (in tag `v0.4.0`). `RegionOne` is usually default region name
in Openstack, but this doesn't justify leaving region to default to this value.

Having region as empty string makes the gophercloud library to pick
first found region in the catalog. This can cause potential backwards
compatibility issues, which are however very unlikely. The release notes will mention that.

Negative impact of having region default to `RegionOne` is that it makes
the plugin to not work in case user relies on region specified in
clouds.yaml.

Signed-off-by: Ondrej Vasko <ondrej.vaskoo@gmail.com>